refactored functions to correct submodules
