+++ ivy/functional/backends/jax/experimental/creation.py
@@ -25,3 +25,34 @@ def triu_indices(
         jnp.triu_indices(n=n_rows, k=k, m=n_cols),
         device=device,
     )
+
+
+def vorbis_window(
+    window_length: JaxArray,
+    *,
+    dtype: Optional[jnp.dtype] = jnp.float32,
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    return jnp.array(
+        [
+            round(
+                math.sin(
+                    (ivy.pi / 2) * (math.sin(ivy.pi * (i) / (window_length * 2)) ** 2)
+                ),
+                8,
+            )
+            for i in range(1, window_length * 2)[0::2]
+        ],
+        dtype=dtype,
+    )
+
+
+def hann_window(
+    window_length: int,
+    periodic: Optional[bool] = True,
+    dtype: Optional[jnp.dtype] = None,
+    *,
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    window_length = window_length + 1 if periodic is True else window_length
+    return jnp.array(jnp.hanning(window_length), dtype=dtype)
