+++ ivy/functional/backends/jax/experimental/layers.py
@@ -7,37 +7,6 @@ import jax.numpy as jnp
 import math
 
 
-def vorbis_window(
-    window_length: JaxArray,
-    *,
-    dtype: Optional[jnp.dtype] = jnp.float32,
-    out: Optional[JaxArray] = None,
-) -> JaxArray:
-    return jnp.array(
-        [
-            round(
-                math.sin(
-                    (ivy.pi / 2) * (math.sin(ivy.pi * (i) / (window_length * 2)) ** 2)
-                ),
-                8,
-            )
-            for i in range(1, window_length * 2)[0::2]
-        ],
-        dtype=dtype,
-    )
-
-
-def hann_window(
-    window_length: int,
-    periodic: Optional[bool] = True,
-    dtype: Optional[jnp.dtype] = None,
-    *,
-    out: Optional[JaxArray] = None,
-) -> JaxArray:
-    window_length = window_length + 1 if periodic is True else window_length
-    return jnp.array(jnp.hanning(window_length), dtype=dtype)
-
-
 def _pool(inputs, init, reduce_fn, window_shape, strides, padding):
 
     if isinstance(strides, int):
@@ -130,83 +99,6 @@ def kaiser_window(
         return jnp.array(jnp.kaiser(M=window_length, beta=beta), dtype=dtype)
     else:
         return jnp.array(jnp.kaiser(M=window_length + 1, beta=beta)[:-1], dtype=dtype)
-
-
-def _flat_array_to_1_dim_array(x):
-    return x.reshape((1,)) if x.shape == () else x
-
-
-def pad(
-    input: JaxArray,
-    pad_width: Union[Sequence[Sequence[int]], JaxArray, int],
-    /,
-    *,
-    mode: Optional[
-        Union[
-            Literal[
-                "constant",
-                "edge",
-                "linear_ramp",
-                "maximum",
-                "mean",
-                "median",
-                "minimum",
-                "reflect",
-                "symmetric",
-                "wrap",
-                "empty",
-            ],
-            Callable,
-        ]
-    ] = "constant",
-    stat_length: Optional[Union[Sequence[Sequence[int]], int]] = None,
-    constant_values: Optional[Union[Sequence[Sequence[Number]], Number]] = 0,
-    end_values: Optional[Union[Sequence[Sequence[Number]], Number]] = 0,
-    reflect_type: Optional[Literal["even", "odd"]] = "even",
-    out: Optional[JaxArray] = None,
-    **kwargs: Optional[Any],
-) -> JaxArray:
-    if callable(mode):
-        return jnp.pad(
-            _flat_array_to_1_dim_array(input),
-            pad_width,
-            mode=mode,
-            **kwargs,
-        )
-    if mode in ["maximum", "mean", "median", "minimum"]:
-        return jnp.pad(
-            _flat_array_to_1_dim_array(input),
-            pad_width,
-            mode=mode,
-            stat_length=stat_length,
-        )
-    elif mode == "constant":
-        return jnp.pad(
-            _flat_array_to_1_dim_array(input),
-            pad_width,
-            mode=mode,
-            constant_values=constant_values,
-        )
-    elif mode == "linear_ramp":
-        return jnp.pad(
-            _flat_array_to_1_dim_array(input),
-            pad_width,
-            mode=mode,
-            end_values=end_values,
-        )
-    elif mode in ["reflect", "symmetric"]:
-        return jnp.pad(
-            _flat_array_to_1_dim_array(input),
-            pad_width,
-            mode=mode,
-            reflect_type=reflect_type,
-        )
-    else:
-        return jnp.pad(
-            _flat_array_to_1_dim_array(input),
-            pad_width,
-            mode=mode,
-        )
 
 
 def max_pool3d(
