+++ ivy/functional/backends/jax/experimental/manipulation.py
@@ -103,3 +103,80 @@ def i0(
     out: Optional[JaxArray] = None,
 ) -> JaxArray:
     return jnp.i0(x)
+
+
+def _flat_array_to_1_dim_array(x):
+    return x.reshape((1,)) if x.shape == () else x
+
+
+def pad(
+    input: JaxArray,
+    pad_width: Union[Sequence[Sequence[int]], JaxArray, int],
+    /,
+    *,
+    mode: Optional[
+        Union[
+            Literal[
+                "constant",
+                "edge",
+                "linear_ramp",
+                "maximum",
+                "mean",
+                "median",
+                "minimum",
+                "reflect",
+                "symmetric",
+                "wrap",
+                "empty",
+            ],
+            Callable,
+        ]
+    ] = "constant",
+    stat_length: Optional[Union[Sequence[Sequence[int]], int]] = None,
+    constant_values: Optional[Union[Sequence[Sequence[Number]], Number]] = 0,
+    end_values: Optional[Union[Sequence[Sequence[Number]], Number]] = 0,
+    reflect_type: Optional[Literal["even", "odd"]] = "even",
+    out: Optional[JaxArray] = None,
+    **kwargs: Optional[Any],
+) -> JaxArray:
+    if callable(mode):
+        return jnp.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            **kwargs,
+        )
+    if mode in ["maximum", "mean", "median", "minimum"]:
+        return jnp.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            stat_length=stat_length,
+        )
+    elif mode == "constant":
+        return jnp.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            constant_values=constant_values,
+        )
+    elif mode == "linear_ramp":
+        return jnp.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            end_values=end_values,
+        )
+    elif mode in ["reflect", "symmetric"]:
+        return jnp.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            reflect_type=reflect_type,
+        )
+    else:
+        return jnp.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+        )
