+++ ivy/functional/backends/numpy/experimental/manipulation.py
@@ -119,3 +119,81 @@ def i0(
 
 
 i0.support_native_out = False
+
+
+def _flat_array_to_1_dim_array(x):
+    return x.reshape((1,)) if x.shape == () else x
+
+
+def pad(
+    input: np.ndarray,
+    pad_width: Union[Sequence[Sequence[int]], np.ndarray, int],
+    /,
+    *,
+    mode: Optional[
+        Union[
+            Literal[
+                "constant",
+                "edge",
+                "linear_ramp",
+                "maximum",
+                "mean",
+                "median",
+                "minimum",
+                "reflect",
+                "symmetric",
+                "wrap",
+                "empty",
+            ],
+            Callable,
+        ]
+    ] = "constant",
+    stat_length: Optional[Union[Sequence[Sequence[int]], int]] = None,
+    constant_values: Optional[Union[Sequence[Sequence[Number]], Number]] = 0,
+    end_values: Optional[Union[Sequence[Sequence[Number]], Number]] = 0,
+    reflect_type: Optional[Literal["even", "odd"]] = "even",
+    out: Optional[np.ndarray] = None,
+    **kwargs: Optional[Any],
+) -> np.ndarray:
+    if callable(mode):
+        return np.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            **kwargs,
+        )
+    if mode in ["maximum", "mean", "median", "minimum"]:
+        return np.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            stat_length=stat_length,
+        )
+    elif mode == "constant":
+        return np.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            constant_values=constant_values,
+        )
+    elif mode == "linear_ramp":
+        return np.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            end_values=end_values,
+        )
+    elif mode in ["reflect", "symmetric"]:
+        return np.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+            reflect_type=reflect_type,
+        )
+    else:
+        return np.pad(
+            _flat_array_to_1_dim_array(input),
+            pad_width,
+            mode=mode,
+        )
+
