+++ ivy/functional/backends/torch/experimental/layers.py
@@ -9,60 +9,6 @@ from ivy.func_wrapper import with_unsupported_dtypes
 from . import backend_version
 
 
-def flatten(
-    x: torch.Tensor,
-    /,
-    *,
-    start_dim: Optional[int] = 0,
-    end_dim: Optional[int] = -1,
-    out: Optional[torch.Tensor] = None,
-) -> torch.Tensor:
-    return torch.flatten(x, start_dim=start_dim, end_dim=end_dim)
-
-
-def vorbis_window(
-    window_length: torch.tensor,
-    *,
-    dtype: Optional[torch.dtype] = torch.float32,
-    out: Optional[torch.tensor] = None,
-) -> torch.tensor:
-    return torch.tensor(
-        [
-            round(
-                math.sin(
-                    (ivy.pi / 2) * (math.sin(ivy.pi * (i) / (window_length * 2)) ** 2)
-                ),
-                8,
-            )
-            for i in range(1, window_length * 2)[0::2]
-        ],
-        dtype=dtype,
-    )
-
-
-vorbis_window.support_native_out = False
-
-
-def hann_window(
-    window_length: int,
-    periodic: Optional[bool] = True,
-    dtype: Optional[torch.dtype] = None,
-    *,
-    out: Optional[torch.tensor] = None,
-) -> torch.tensor:
-    return torch.hann_window(
-        window_length,
-        periodic=periodic,
-        dtype=dtype,
-        layout=torch.strided,
-        device=None,
-        requires_grad=None,
-    )
-
-
-hann_window.support_native_out = False
-
-
 @with_unsupported_dtypes(
     {
         "1.11.0 and below": (
@@ -142,47 +88,6 @@ def max_pool1d(
     if data_format == "NWC":
         res = res.permute(0, 2, 1)
     return res
-
-
-def kaiser_window(
-    window_length: int,
-    periodic: bool = True,
-    beta: float = 12.0,
-    *,
-    dtype: Optional[torch.dtype] = None,
-    out: Optional[torch.Tensor] = None,
-) -> torch.Tensor:
-    return torch.kaiser_window(
-        window_length,
-        periodic,
-        beta,
-        dtype=dtype,
-        layout=torch.strided,
-        device=None,
-        requires_grad=False,
-    )
-
-
-def hamming_window(
-    window_length: int,
-    /,
-    *,
-    periodic: Optional[bool] = True,
-    alpha: Optional[float] = 0.54,
-    beta: Optional[float] = 0.46,
-    dtype: Optional[torch.dtype] = None,
-    out: Optional[torch.Tensor] = None,
-) -> torch.Tensor:
-    return torch.hamming_window(
-        window_length,
-        periodic=periodic,
-        alpha=alpha,
-        beta=beta,
-        dtype=dtype,
-        layout=torch.strided,
-        device=None,
-        requires_grad=False,
-    )
 
 
 @with_unsupported_dtypes(
